requirements.txt
beautifulsoup4
requests
pandas
streamlit
streamlit-autorefresh
streamlit
opencv-python
numpy
pillow
matplotlib
streamlit
opencv-python
numpy
pillow
matplotlib
streamlit-extras
